requirements.txt
python-telegram-bot
PyMuPDF
python-docx
requests
python-telegram-bot
PyMuPDF
python-docx
requests
dotenv
pyspellchecker